feat: retposto enhancements - help on no-args, email in output, better errors
Fixes issue #17:

- Split monolithic cli.py (1418 lines) into cli/ package (8 files, all <500 lines)
- Change no_args_is_help=False to True on retposto, konton (consistent with rest)
- retposto_preni: show email instead of UUID; filter accounts with passwords
  before counting; add UUID prefix resolution for --account flag
- _report_sync: accept account_label for context in error messages
- imap.py: include username@host:port in connection errors
- imap.py: add [email] context to concurrent sync errors
- Fix pre-existing test bugs: DB isolation fixture, sync_folder signature,
  malfermi -> malfari typo, missing kategorio_app
