+++ src/A_lien/cli/kategorio.py
@@ -1,0 +1,101 @@
+"""Kategorio sub-typer — category management.
+
+Commands: ls, aldoni, forigi
+Registered on kontakto typer from kontakto.py.
+"""
+
+from __future__ import annotations
+
+import typer
+
+from A import error, info, tr_multi
+from A_lien.service import get_kontakto_service
+
+kategorio_app = typer.Typer(
+    name="kategorio",
+    help=tr_multi(
+        "Administri kategoriojn.",
+        "Manage categories.",
+        "Gérer les catégories.",
+    ),
+    no_args_is_help=True,
+    context_settings={"help_option_names": ["-h", "--help", "--helpo"]},
+)
+
+
+@kategorio_app.command("ls")
+def kategorio_ls() -> None:
+    """List categories."""
+    service = get_kontakto_service()
+    cats = service.list_categories()
+    if not cats:
+        info(tr_multi(
+            "Neniuj kategorioj.",
+            "No categories.",
+            "Aucune catégorie.",
+        ))
+        return
+    for c in cats:
+        color = f" ({c['koloro']})" if c.get("koloro") else ""
+        info(f"  {c['uuid'][:8]}  {c['nomo']}{color}")
+
+
+@kategorio_app.command("aldoni")
+def kategorio_aldoni(
+    nomo: str = typer.Argument(
+        ...,
+        help=tr_multi("Kategorinomo", "Category name", "Nom de catégorie"),
+    ),
+    koloro: str = typer.Option(
+        "", "--koloro", "-k",
+        help=tr_multi("Kolorkodo", "Color code", "Code couleur"),
+    ),
+) -> None:
+    """Add a new category."""
+    service = get_kontakto_service()
+    try:
+        cat = service.create_category(nomo, koloro)
+        info(tr_multi(
+            f"Kategorio kreita: {cat['uuid'][:8]} {nomo}",
+            f"Category created: {cat['uuid'][:8]} {nomo}",
+            f"Catégorie créée: {cat['uuid'][:8]} {nomo}",
+        ))
+    except Exception as e:
+        error(tr_multi(
+            f"Eraro: {e}",
+            f"Error: {e}",
+            f"Erreur: {e}",
+        ))
+        raise typer.Exit(1)
+
+
+@kategorio_app.command("forigi")
+def kategorio_forigi(
+    uuid: str = typer.Argument(
+        ...,
+        help=tr_multi("Kategorio UUID", "Category UUID", "UUID catégorie"),
+    ),
+) -> None:
+    """Delete a category."""
+    service = get_kontakto_service()
+    if service.delete_category(uuid):
+        info(tr_multi(
+            f"Kategorio forigita: {uuid[:8]}",
+            f"Category deleted: {uuid[:8]}",
+            f"Catégorie supprimée: {uuid[:8]}",
+        ))
+    else:
+        error(tr_multi(
+            f"Kategorio ne trovita: {uuid}",
+            f"Category not found: {uuid}",
+            f"Catégorie non trouvée: {uuid}",
+        ))
+        raise typer.Exit(1)
+
+
+__all__ = [
+    "kategorio_app",
+    "kategorio_ls",
+    "kategorio_aldoni",
+    "kategorio_forigi",
+]
